backend_olleh/pyproject.toml
[project]
name = "backend-olleh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-guardian>=3.2.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-spectacular>=0.29.0",
    "markdown>=3.10.1",
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
    "ty>=0.0.14",
]
